rel_path

StatisticalRethinking.rel_path — Method

rel_path

Relative path using the StatisticalRethinking src/ directory. Copied from DynamicHMCExamples.jl

Example to get access to the data subdirectory

rel_path("..", "data")

link

StatisticalRethinking.link — Function

link

Compute the link function for standadized vars.

link(dfa, vars, xrange)

Extended help

Required arguments

* `df::DataFrame`               : Chain samples converted to a DataFrame
* `vars::Vector{Symbol}`        : Variables in DataFrame (2 variables)
* `xrange::range`               : Range over which link values are computed

Optional arguments

* `xbar::Float64`               : Mean value of observed predictor
* `ybar::Float64`               : Mean value of observed outcome (requires xbar argument)

Return values

* `result`                      : Vector of link values

rescale

StatisticalRethinking.rescale — Function

rescale

Rescale a vector or an array to "un-standardize", the opposite of scale!()

rescale(x, xbar, xstd)

Extended help

Required arguments

* `x::Vector{Float64}`                 : Vector to be rescaled
* `xbar`                               : Mean value for rescaling
* `xstd`                               : Std for rescaling

Return values

* `result::AbstractVector`            : Rescaled vector

quap

StatisticalRethinking.quap — Method

quap

Quadratic approximation of a posterior distribution

Method

quap(df) 

Required arguments

* `df::DataFrame`               : Dataframe generated from a chain

Return values

* `result::NamedTuple`          : NamedTuple representing the quadratic approximation

To convert to a Dict use:

dct = Dict(pairs(result))

Example

# Run stan_sample() on a SampleModel

if success(rc)
	
	df = read_samples(sm; output_format=:dataframe)
	q = quap(df)

end

sample

StatsBase.sample — Function

sample([rng], a, [wv::AbstractWeights])

Select a single random element of a. Sampling probabilities are proportional to the weights given in wv, if provided.

Optionally specify a random number generator rng as the first argument (defaults to Random.GLOBAL_RNG).

sample([rng], a, [wv::AbstractWeights], n::Integer; replace=true, ordered=false)

Select a random, optionally weighted sample of size n from an array a using a polyalgorithm. Sampling probabilities are proportional to the weights given in wv, if provided. replace dictates whether sampling is performed with replacement and order dictates whether an ordered sample, also called a sequential sample, should be taken.

Optionally specify a random number generator rng as the first argument (defaults to Random.GLOBAL_RNG).

sample([rng], a, [wv::AbstractWeights], dims::Dims; replace=true, ordered=false)

Select a random, optionally weighted sample from an array a specifying the dimensions dims of the output array. Sampling probabilities are proportional to the weights given in wv, if provided. replace dictates whether sampling is performed with replacement and order dictates whether an ordered sample, also called a sequential sample, should be taken.

Optionally specify a random number generator rng as the first argument (defaults to Random.GLOBAL_RNG).

sample([rng], wv::AbstractWeights)

Select a single random integer in 1:length(wv) with probabilities proportional to the weights given in wv.

Optionally specify a random number generator rng as the first argument (defaults to Random.GLOBAL_RNG).

sample

Sample rows from a DataFrame

Method

sample(df, n; replace, ordered) 

Required arguments

* `df::DataFrame`               : DataFrame
* `n::Int`                      : Number of samples

Optional argument

* `rng::AbstractRNG`            : Random number generator
* `replace::Bool=true`          : Sample with replace 
* `ordered::Bool=false`         : Sort sample 

Return values

* `result`                      : Array of samples

hpdi

Missing docstring.

Missing docstring for hpdi. Check Documenter's build log for details.

plotcoef

Missing docstring.

Missing docstring for plotcoef. Check Documenter's build log for details.

pairsplot

Missing docstring.

Missing docstring for pairsplot. Check Documenter's build log for details.

plotbounds

Missing docstring.

Missing docstring for plotbounds. Check Documenter's build log for details.

convert_a3d

StatisticalRethinking.convert_a3d — Method

convert_a3d

Convert the output file(s) created by cmdstan to a single DataFrame.

StatisticalRethinking.convert_a3d — Method

convert_a3d

Convert the output file(s) created by cmdstan to a Vector{DataFrame).