Functions

maximum_a_posteriori

StatisticalRethinking.maximum_a_posteriori — Method.

maximumaposterior

Compute the maximumaposteriori of a model.

Method

maximum_a_posteriori(model, lower_bound, upper_bound)

Required arguments

* `model::Turing model`
* `lower_bound::Float64`
* `upper_bound::Float64`

Return values

* `result`                       : Maximum_a_posterior vector

Examples

See 02/clip_08t.jl

source

rel_path

StatisticalRethinking.rel_path — Method.

rel_path

Relative path using the StatisticalRethinking src/ directory. Copied from DynamicHMCExamples.jl

Example to get access to the data subdirectory

rel_path("..", "data")

source

ScriptEntry

StatisticalRethinking.ScriptEntry — Type.

ScriptEntry

Define processing requirements for chapter scripts

Constructor

scriptentry(scriptfile;; nb=true, exe=true, doc=true)

Required arguments

* `scriptfile::AbstractString`        : Script file

Optional arguments

* `nb::Bool`      : Generate a notebook version in notebooks directory
* `exe::Bool`     : Execute the notebook (for testing or documentation purposes)
* `doc::Bool`     : Insert documention into Github pages

If exe = false and doc = true it is assumed the appropriate .md files have been manually created and stored in docs/src/nn/... (Travis will terminate if runs take too long).

source

scriptentry

StatisticalRethinking.scriptentry — Method.

scriptentry

Constructor for ScriptEntry objects.

Constructor

scriptentry(scriptfile;; nb=true, exe=true, doc=true)

Required arguments

* `scriptfile::AbstractString`        : Script file

Optional arguments

* `nb::Bool`      : Generate a notebook version in notebooks directory
* `exe::Bool`     : Execute the notebook (for testing or documentation purposes)
* `doc::Bool`     : Insert documention into Github pages

If exe = false and doc = true it is assumed the appropriate .md files have been manually created and stored in docs/src/nn/... (Travis will terminate if runs take too long).

source

link

StatisticalRethinking.link — Method.

link

Compute the link function

Method

link(xrange, chain, vars, xbar) 

Required arguments

* `xrange::Turing model`  : Range over which link values are computed
* `chain::Float64`             : Chain samples used
* `vars::Float64`               : Variables in chain used
* `xbar::Float64`               : Mean value of observed predictor

Return values

* `result`                       : Vector of link values

source