sr_path

StatisticalRethinking.sr_path — Method

sr_path

Relative path using the StatisticalRethinking src/ directory.

Example to get access to the data subdirectory

sr_path("..", "data")

Note that in the projects, e.g. StatisticalRethinkingStan.jl and StatisticalRethinkingTuring.jl, the DrWatson approach is a better choics, i.e: sr_datadir(filename)

sr_datadir

StatisticalRethinking.sr_datadir — Method

sr_datadir

Relative path using the StatisticalRethinking src/ directory.

Example to access Howell1.csv in StatisticalRethinking:

df = CSV.read(sr_datadir("Howell1.csv"), DataFrame)

link

StatisticalRethinking.link — Method

link

Compute the link function for standardized variables.

link(dfa, vars, xrange)

Extended help

Required arguments

* `df::DataFrame`                      : Chain samples converted to a DataFrame
* `vars::Vector{Symbol}`               : Variables in DataFrame (2 variables)
* `xrange::range`                      : Range over which link values are computed

Optional arguments

* `xbar::Float64`                      : Mean value of observed predictor
* `ybar::Float64`                      : Mean value of observed outcome (requires xbar argument)

Return values

* `result`                             : Vector of link values

rescale

StatisticalRethinking.rescale — Method

rescale

Rescale a vector to "un-standardize", the opposite of scale!().

rescale(x, xbar, xstd)

Extended help

Required arguments

* `x::Vector{Float64}`                 : Vector to be rescaled
* `xbar`                               : Mean value for rescaling
* `xstd`                               : Std for rescaling

Return values

* `result::AbstractVector`             : Rescaled vector

sample

StatsBase.sample — Method

sample

Sample rows from a DataFrame

Method

sample(df, n; replace, ordered) 

Required arguments

* `df::DataFrame`                      : DataFrame
* `n::Int`                             : Number of samples

Optional argument

* `rng::AbstractRNG`                   : Random number generator
* `replace::Bool=true`                 : Sample with replace 
* `ordered::Bool=false`                : Sort sample 

Return values

* `result`                             : Array of samples

hpdi

StatisticalRethinking.hpdi — Method

hpdi

Compute high density region.

hpdi(x; alpha)

Derived from hpd in MCMCChains.jl.

By default alpha=0.11 for a 2-sided tail area of p < 0.055% and p > 0.945%.

pairsplot

StatisticalRethinking.pairsplot — Method

pairsplot

A simple version of StatsPlots' cornerplot.

pairsplot(df, vars)

Required arguments

* `df::DataFrame`                      : DataFrame containing the variables (as columns)
* `vars::Vector{Symbol}`               : Vector of variables to include in plot

meanlowerupper

StatisticalRethinking.meanlowerupper — Function

meanlowerupper

Compute a NamedTuple with means, lower and upper PI values.

meanlowerupper(data)
meanlowerupper(data, PI)

plotbounds

StatisticalRethinking.plotbounds — Method

plotbounds

Plot regression line and intervals based on Stan samples of coeffficients.

plotbounds(df, xvar, yvar, dfs, linkvars; fnc, fig, bounds, title, xlab, ylab, alpha, colors, stepsize, rescale_axis, kwargs...)

Required arguments

* `df::DataFrame`                      : DataFrame with observed variables and scaled variables
* `xvar::Symbol`                       : X variable in df
* `yvar::Symbol`                       : Y variable in df
* `dfs::DataFrame`                     : DataFrame with Stan samples
* `linkvars::Vector{Symbol}`           : Initial 2 Symbols are regression coefficients,
                                         3rd - only for :predict bounds indicates σ.

Optional arguments

* `fnc = link`                         : Link function
* `fig::AbstractString=""`             : File to store the plot. If "", a plot is returned
* `bounds::Vector{Symbol}`             : Bounds to display, see below
* `title::AbstractString=""`           : Title for plot
* `xlab::AbstractString=String(xvar)`  : X axis variable label
* `ylab::AbstractString=String(yvar)`  : Y axis variable label
* `alpha::Float64=0.11`                : Interval value, defaults to [0.045, 0.945]
* `colors::Vector{Symbol}`             : Colors for regions, defaults to [:lightgrey, :grey]
* `stepsize::Float64=0.01`             : Stepsize for boundary accuracy 
* `rescale_axis=true`                  : Display using un-standardized scale         

This method is primarily intended to display 2 regions, typically predicted values and the quantile or hpdi region around the mean line. This is specified using the bounds keyword argument, e.g. bounds = [:predicted, :hpdi].

For the prediction interval, a 3rd parameter needs to be present in the Stan sample DataFrame (dfs) containing the σ value. This symbol needs to be added to linkvars, e.g.

linkvars = [:a, :bM, :sigma]

For other options, :quantile and :hpdi, two parameters suffice (typically the itercept and slope parameters).

StatisticalRethinking.plotbounds — Method

plotbounds

Plot regression line and PI interval with alpha=0.11.

plotbounds(df, xvar, yvar, nt, linkvars; fnc, fig, stepsize, rescale_axis, lab, title, kwargs...)

Required arguments

* `df::DataFrame`                      : DataFrame with observed variables and scaled variables
* `xvar::Symbol`                       : X variable in df
* `yvar::Symbol`                       : Y variable in df
* `nt::NamedTuple`                     : NamedTuple with Stan samples
* `linkvars::Vector{Symbol}`           : Initial 2 Symbols are regression coefficients,
                                         3rd - only for :predict bounds indicates σ.

Optional arguments

* `fnc = link`                         : Link function
* `fig::AbstractString=""`             : File to store the plot. If "", a plot is returned
* `stepsize::Float64=0.01`             : Stepsize for boundary accuracy 
* `rescale_axis=true`                  : Display using un-standardized scale         
* `title::AbstractString=""`           : Title for plot
* `lab::AbstractString=""`             : X axis variable label

This method is intended to the PI region around the mean line.

The 2nd symbol in linkvars needs to match the matrix

linkvars = [:a, :bM]

For other options, :quantile and :hpdi, two parameters suffice (typically the itercept and slope parameters).

plotlines

StatisticalRethinking.plotlines — Function

plotlines

Plot regression line.

plotlines(df, xvar, yvar, nt, linkvars)
plotlines(df, xvar, yvar, nt, linkvars, fig; fnc, stepsize, rescale_axis, lab, title, kwargs...)

Required arguments

* `df::DataFrame`                      : DataFrame with observed variables and scaled variables
* `xvar::Symbol`                       : X variable in df
* `yvar::Symbol`                       : Y variable in df
* `nt::NamedTuple`                     : NamedTuple with Stan samples
* `linkvars::Vector{Symbol}`           : Initial 2 Symbols are regression coefficients,
                                         3rd - only for :predict bounds indicates σ.

Optional arguments

* `fnc = link`                         : Link function
* `fig::AbstractString=""`             : File to store the plot. If "", a plot is returned
* `stepsize::Float64=0.01`             : Stepsize for boundary accuracy 
* `rescale_axis=true`                  : Display using un-standardized scale         
* `title::AbstractString=""`           : Title for plot
* `lab::AbstractString=""`             : X axis variable label

This method is primarily intended to show multiple mean lines, e.g. in case of LOO.

The 2nd symbol in linkvars needs to match the matrix generated by fnc

linkvars = [:a, :bM]

compare

StatisticalRethinking.compare — Method

compare

Compare waic and psis values for models.

compare(m, ; mnames)

Required arguments

* `models`                             : Vector of logprob matrices
* `criterium`                          : Either ::Val{:waic} or ::Val{:psis}

Optional argument

* `mnames::Vector{Symbol}`             : Vector of model names

Return values

* `df`                                 : DataFrame with statistics

create_observation_matrix

StatisticalRethinking.create_observation_matrix — Method

pairsplot

Create a polynomial observation matrix.

create_observation_matrix(x, k)

r2_is_bad

StatisticalRethinking.r2_is_bad — Method

r2isbad

Compute R^2 values.

r2_is_bad(model, df)

var2

StatisticalRethinking.var2 — Method

var2

Variance without n-1 correction.

var2(x)

simulate

StatisticalRethinking.simulate — Method

simulate

Used for counterfactual simulations.

simulate(df, coefs, var_seq)

Required arguments

* `df`                                 : DataFrame with coefficient samples
* `coefs`                              : Vector of coefficients
* `var_seq`                            : Input values for simulated effect

Return values

* `m_sim::NamedTuple`                  : Array with predictions

StatisticalRethinking.simulate — Method

simulate

Counterfactual predictions after manipulating a variable.

simulate(df, coefs, var_seq, coefs_ext)

Required arguments

* `df`                                 : DataFrame with coefficient samples
* `coefs`                              : Vector of coefficients
* `var_seq`                            : Input values for simulated effect
* `ext_coefs`                          : Vector of simulated variable coefficients

Return values

* `(m_sim, d_sim)`                     : Arrays with predictions