sr_path

StatisticalRethinking.sr_path — Method

sr_path

Relative path using the StatisticalRethinking src/ directory.

Example to get access to the data subdirectory

sr_path("..", "data")

Note that in the projects, e.g. StatisticalRethinkingStan.jl and StatisticalRethinkingTuring.jl, the DrWatson approach is a better choics, i.e: sr_datadir(filename)

sr_datadir

StatisticalRethinking.sr_datadir — Method

sr_datadir

Relative path using the StatisticalRethinking src/ directory.

Example to access Howell1.csv in StatisticalRethinking:

df = CSV.read(sr_datadir("Howell1.csv"), DataFrame)

link

StatisticalRethinking.link — Method

link

Compute the link function for standardized variables.

link(dfa, vars, xrange)

Extended help

Required arguments

* `df::DataFrame`                      : Chain samples converted to a DataFrame
* `vars::Vector{Symbol}`               : Variables in DataFrame (2 variables)
* `xrange::range`                      : Range over which link values are computed

Optional arguments

* `xbar::Float64`                      : Mean value of observed predictor
* `ybar::Float64`                      : Mean value of observed outcome (requires xbar argument)

Return values

* `result`                             : Vector of link values

rescale

StatisticalRethinking.rescale — Method

rescale

Rescale a vector to "un-standardize", the opposite of scale!().

rescale(x, xbar, xstd)

Extended help

Required arguments

* `x::Vector{Float64}`                 : Vector to be rescaled
* `xbar`                               : Mean value for rescaling
* `xstd`                               : Std for rescaling

Return values

* `result::AbstractVector`             : Rescaled vector

sample

StatsBase.sample — Method

sample

Sample rows from a DataFrame

Method

sample(df, n; replace, ordered) 

Required arguments

* `df::DataFrame`                      : DataFrame
* `n::Int`                             : Number of samples

Optional argument

* `rng::AbstractRNG`                   : Random number generator
* `replace::Bool=true`                 : Sample with replace 
* `ordered::Bool=false`                : Sort sample 

Return values

* `result`                             : Array of samples

hpdi

StatisticalRethinking.hpdi — Method

hpdi

Compute high density region.

hpdi(x; alpha)

Derived from hpd in MCMCChains.jl.

By default alpha=0.11 for a 2-sided tail area of p < 0.055% and p > 0.945%.

pairsplot

Missing docstring.

Missing docstring for pairsplot(df::DataFrame, vars::Vector{Symbol}, fig::AbstractString). Check Documenter's build log for details.

plotbounds

StatisticalRethinking.plotbounds — Method

plotbounds

Plot regression line and intervals based on Stan samples of coeffficients.

plotbounds(df, xvar, yvar, dfs, linkvars; fig, bounds, title, xlab, ylab, alpha, colors, stepsize, rescale_axis, kwargs...)

Required arguments

* `df::DataFrame`                      : DataFrame with observed variables and scaled variables
* `xvar::Symbol`                       : X variable in df
* `yvar::Symbol`                       : Y variable in df
* `dfs::DataFrame`                     : DataFrame with Stan samples
* `linkvars::Vector{Symbol}`           : Initial 2 Symbols are regression coefficients,
                                         3rd - only for :predict bounds indicates σ.

Optional arguments

* `fig::AbstractString=""`             : File to store the plot. If "", a plot is returned
* `bounds::Vector{Symbol}`             : Bounds to display, see below
* `title::AbstractString=""`           : Title for plot
* `title::AbstractString=String(xvar)` : X axis variable label
* `title::AbstractString=String(yvar)` : Y axis variable label
* `alpha::Float64=0.11`                : Interval value, defaults to [0.045, 0.945]
* `colors::Vector{Symbol}`             : Colors for regions, defaults to [:lightgrey, :grey]
* `stepsize::Float64=0.01`             : Stepsize for boundary accuracy 
* `rescale_axis=true`                  : Display using un-standardized scale         

This method is primarily intended to display 2 regions, typically predicted values and the quantile or hpdi region around the mean line. This is specified using the bounds keyword argument, e.g. bounds = [:predicted, :hpdi].

For the prediction interval, a 3rd parameter needs to be present in the Stan sample DataFrame (dfs) containing the σ value. This symbol needs to be added to linkvars, e.g.

linkvars = [:a, :bM, :sigma]

For other options, :quantile and :hpdi, two parameters suffice (typically the itercept and slope parameters).

simulate

StatisticalRethinking.simulate — Method

simulate

Used for counterfactual simulations.

simulate(df, coefs, var_seq)

Required arguments

* `df`                                 : DataFrame with coefficient samples
* `coefs`                              : Vector of coefficients
* `var_seq`                            : Input values for simulated effect

Return values

* `m_sim::NamedTuple`                  : Array with predictions

StatisticalRethinking.simulate — Method

simulate

Counterfactual predictions after manipulating a variable.

simulate(df, coefs, var_seq, coefs_ext)

Required arguments

* `df`                                 : DataFrame with coefficient samples
* `coefs`                              : Vector of coefficients
* `var_seq`                            : Input values for simulated effect
* `ext_coefs`                          : Vector of simulated variable coefficients

Return values

* `(m_sim, d_sim)`                     : Arrays with predictions