rel_path

StatisticalRethinking.rel_path — Method

rel_path

Relative path using the StatisticalRethinking src/ directory.

Example to get access to the data subdirectory

rel_path("..", "data")

Note that in the projects, e.g. StatisticalRethinkingStan.jl and StatisticalRethinkingTuring.jl, the DrWatson approach is a better choics, i.e: sr_datadir(filename)

link

StatisticalRethinking.link — Method

link

Compute the link function for standardized variables.

link(dfa, vars, xrange)

Extended help

Required arguments

* `df::DataFrame`                      : Chain samples converted to a DataFrame
* `vars::Vector{Symbol}`               : Variables in DataFrame (2 variables)
* `xrange::range`                      : Range over which link values are computed

Optional arguments

* `xbar::Float64`                      : Mean value of observed predictor
* `ybar::Float64`                      : Mean value of observed outcome (requires xbar argument)

Return values

* `result`                             : Vector of link values

rescale

StatisticalRethinking.rescale — Method

rescale

Rescale a vector to "un-standardize", the opposite of scale!().

rescale(x, xbar, xstd)

Extended help

Required arguments

* `x::Vector{Float64}`                 : Vector to be rescaled
* `xbar`                               : Mean value for rescaling
* `xstd`                               : Std for rescaling

Return values

* `result::AbstractVector`             : Rescaled vector

quap

StatisticalRethinking.quap — Method

quap

Quadratic approximation of a posterior distribution

Method

quap(df) 

Required arguments

* `df::DataFrame`                      : Dataframe generated from samples (chains)

Return values

* `result::NamedTuple`                 : NamedTuple representing the quadratic approximation

To convert to a Dict use:

dct = Dict(pairs(result))

Example

# Run stan_sample() on a SampleModel

if success(rc)
	
	df = read_samples(sm; output_format=:dataframe)
	q = quap(df)

end

sample

StatsBase.sample — Method

sample

Sample rows from a DataFrame

Method

sample(df, n; replace, ordered) 

Required arguments

* `df::DataFrame`                      : DataFrame
* `n::Int`                             : Number of samples

Optional argument

* `rng::AbstractRNG`                   : Random number generator
* `replace::Bool=true`                 : Sample with replace 
* `ordered::Bool=false`                : Sort sample 

Return values

* `result`                             : Array of samples

hpdi

StatisticalRethinking.hpdi — Method

hpdi

Compute high density region.

hpdi(x; alpha)

Derived from hpd in MCMCChains.jl.

Alpha represents the 2-sided tail area, i.e. p < 0.055% and p > 0.945%.

plotcoef

StatisticalRethinking.plotcoef — Function

plotcoef

Multiple regression coefficient plot.

plotcoef(models, pars, fig)
plotcoef(models, pars, fig, title)
plotcoef(models, pars, fig, title, func)

Required arguments

* `models`                             : Vector of `SampleModel`s to compare
* `pars`                               : Vector of parameters to include in comparison
* `fig`                                : File to store the produce plot

Optional arguments

* `title=""`                           : Title for plot
* `func=nothing`                       : Optional funtion to apply to sample dataframe

Currently the only function available is quap.

The function will be called with a single argument, a dataframe constructed from all samples in all chains in the SampleModels.

It return a Partcles type NamedTuple. e.g.:

(a = 0.000527 ± 0.1, bM = -0.0628 ± 0.16, bA = -0.608 ± 0.16, sigma = 0.828 ± 0.089)

An example can be found in scipts/05/clip-13.jl.

Return values

* `result::NamedTuple`                 : Vector{NamedTuple} of estimates (Particles or Quap)

StatisticalRethinking.plotcoef — Function

plotcoef

Multiple regression coefficient plot.

plotcoef(model, pars, fig)
plotcoef(model, pars, fig, title)
plotcoef(model, pars, fig, title, func)

Required arguments

* `model`                              : SampleModel to display
* `pars`                               : Vector of parameters to include in comparison
* `fig`                                : File to store the produced plot

Optional arguments

* `title=""`                           : Title for plot
* `func=nothing`                       : Optional funtion to apply to sample dataframe

Currently the only function available is quap.

The function will be called with a single argument, a dataframe constructed from all samples in all chains in the SampleModel.

It returns a Particles type NamedTuple. e.g.:

(a = 0.000527 ± 0.1, bM = -0.0628 ± 0.16, bA = -0.608 ± 0.16, sigma = 0.828 ± 0.089)

An examples can be found in scipts/06/clip-12-05.jl.

Return values

* `result::NamedTuple`                 : NamedTuple of estimates (Particles or Quap)

pairsplot

StatisticalRethinking.pairsplot — Method

pairsplot

A simple version of StatsPlots' cornerplot.

pairsplot(df, vars, fig)

Required arguments

* `df::DataFrame`                      : DataFrame containing the variables (as columns)
* `vars::Vector{Symbol}`               : Vector of variables to include in plot
* `fig::AbstractString`                : File to store the produced plot

plotbounds

StatisticalRethinking.plotbounds — Method

plotbounds

Plot regression line and intervals based on Stan samples of coeffficients.

plotbounds(df, xvar, yvar, dfs, linkvars; fig, bounds, title, xlab, ylab, alpha, colors, stepsize, rescale_axis)

Required arguments

* `df::DataFrame`                      : DataFrame with observed variables and scaled variables
* `xvar::Symbol`                       : X variable in df
* `yvar::Symbol`                       : Y variable in df
* `dfs::DataFrame`                     : DataFrame with Stan samples
* `linkvars::Vector{Symbol}`           : Initial 2 Symbols are regression coefficients,
                                         3rd - only for :predict bounds indicates σ.

Optional arguments

* `fig::AbstractString=""`             : File to store the plot. If "", a plot is returned
* `bounds::Vector{Symbol}`             : Bounds to display, see below
* `title::AbstractString=""`           : Title for plot
* `title::AbstractString=String(xvar)` : X axis variable label
* `title::AbstractString=String(yvar)` : Y axis variable label
* `alpha::Float64=0.11`                : Interval value, defaults to [0.045, 0.945]
* `colors::Vector{Symbol}`             : Colors for regions, defaults to [:lightgrey, :grey]
* `stepsize::Float64=0.01`             : Stepsize for boundary accuracy 
* `rescale_axis=true`                  : Display using un-standardized scale         

This method is primarily intended to display 2 regions, typically predicted values and the quantile or hpdi region around the mean line. This is specified using the bounds keyword argument, e.g. bounds = [:predicted, :hpdi].

For the prediction interval, a 3rd parameter needs to be present in the Stan sample DataFrame (dfs) containing the σ value. This symbol needs to be added to linkvars, e.g.

linkvars = [:a, :bM, :sigma]

For other options, :quantile and :hpdi, two parameters suffice (typically the itercept and slope parameters).

simulate

StatisticalRethinking.simulate — Method

simulate

Used for counterfactual simulations.

simulate(df, coefs, var_seq)

Required arguments

* `df`                                 : DataFrame with coefficient samples
* `coefs`                              : Vector of coefficients
* `var_seq`                            : Input values for simulated effect

Return values

* `m_sim::NamedTuple`                  : Array with predictions

StatisticalRethinking.simulate — Method

plotcoef

Counterfactual predictions after manipulating a variable.

simulate(df, coefs, var_seq, coefs_ext)

Required arguments

* `df`                                 : DataFrame with coefficient samples
* `coefs`                              : Vector of coefficients
* `var_seq`                            : Input values for simulated effect
* `ext_coefs`                          : Vector of simulated variable coefficients

Return values

* `(m_sim, d_sim)`                     : Arrays with predictions

##convert_a3d

StatisticalRethinking.convert_a3d — Method

convert_a3d

convert_a3d(a3d_array, cnames, _)

Convert the output file(s) created by stan_sample() to a single DataFrame.

StatisticalRethinking.convert_a3d — Method

convert_a3d

convert_a3d(a3d_array, cnames, _)

Convert the output file(s) created by stan_sample() to a Vector{DataFrame).