Functions

link

StatisticalRethinking.link — Method.

link

Compute the link function

Method

link(xrange, chain, vars, xbar) 

Required arguments

* `xrange::Turing model`  : Range over which link values are computed
* `chain::Float64`             : Chain samples used
* `vars::Float64`               : Variables in chain used
* `xbar::Float64`               : Mean value of observed predictor

Return values

* `result`                       : Vector of link values

source

rel_path

StatisticalRethinking.rel_path — Method.

rel_path

Relative path using the StatisticalRethinking src/ directory. Copied from DynamicHMCExamples.jl

Example to get access to the data subdirectory

rel_path("..", "data")

source

generate

StatisticalRethinking.generate — Method.

generate

Utility function to generate all notebooks and chapters from scripts in the scripts directory.

Method

generate(sd = script_dict)

Required arguments

None, all notebooks/.. and chapters/.. files are regenerated.

source

StatisticalRethinking.generate — Method.

generate

Generate notebooks and scripts in a single chapter.

Method

generate(chapter::AbstractString)

Required arguments

Generate notebooks and scripts in a single chapter, e.g. generate("04")

source

StatisticalRethinking.generate — Method.

generate

Generate a single notebook and script

Method

generate(chapter::AbstractString, file::AbstractString)

Required arguments

Generate notebook and script file in chapter, e.g. generate("04", "m4.1d.jl") or generate("04/m4.1d.jl")

source

ScriptEntry

StatisticalRethinking.ScriptEntry — Type.

ScriptEntry

Define processing requirements for chapter scripts

Constructor

scriptentry(scriptfile;; nb=true, exe=true, doc=true)

Required arguments

* `scriptfile::AbstractString`        : Script file

Optional arguments

* `nb::Bool`      : Generate a notebook version in notebooks directory
* `exe::Bool`     : Execute the notebook (for testing or documentation purposes)
* `doc::Bool`     : Insert documention into Github pages

If exe = false and doc = true it is assumed the appropriate .md files have been manually created and stored in docs/src/nn/... (Travis will terminate if runs take too long).

source

scriptentry

StatisticalRethinking.scriptentry — Method.

scriptentry

Constructor for ScriptEntry objects.

Constructor

scriptentry(scriptfile;; nb=true, exe=true, doc=true)

Required arguments

* `scriptfile::AbstractString`        : Script file

Optional arguments

* `nb::Bool`      : Generate a notebook version in notebooks directory
* `exe::Bool`     : Execute the notebook (for testing or documentation purposes)
* `doc::Bool`     : Insert documention into Github pages

If exe = false and doc = true it is assumed the appropriate .md files have been manually created and stored in docs/src/nn/... (Travis will terminate if runs take too long).

source